discord-ai-bot/requirements.txt
discord.py
aiohttp
requests
tinydb
python-dotenv
pyttsx3
